initial working tree remote code completion
